requirements.txt
streamlit
# requirements.txt

# OCR and Image Processing
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.9.0.80
numpy==1.26.0

google-generativeai

